RPX plugin for Hotaru 1.0 (the last of 41 plugins!)

+++ content/plugins/rpx/templates/rpx_register.php
@@ -25,57 +25,51 @@
  */
  
 if ($h->cage->post->getAlpha('users_type') == 'register') {
-    $username_check = $h->cage->post->testUsername('username');
-    $password_check = "";
-    $password2_check = "";
-    $email_check = $h->cage->post->testEmail('email');    
-} else {
-    $username_check = "";
-    $password_check = "";
-    $password2_check = "";
-    $email_check = "";
+    $h->vars['rpx_profile']['preferredUsername'] = $h->cage->post->testUsername('username');
+    $h->vars['rpx_profile']['email'] = $h->cage->post->testEmail('email');
+    $h->vars['rpx_profile']['identifier'] = $h->cage->post->testUri('identifier');
 }
+
 ?>
-    <h2><?php echo $h->lang["user_signin_register"]; ?></h2>
+
+<h2><?php echo $h->lang["user_signin_register"]; ?></h2>
+
+<?php if (!isset($h->vars['rpx_already_exists'])) { ?>
     
     <?php echo $h->showMessages(); ?>
-    
-    <?php $h->pluginHook('user_signin_register_pre_register_form'); ?>
         
     <div class='user_login_reg'>
-    <?php echo $h->lang["user_signin_register_instructions"]; ?>
+    <?php echo $h->lang["rpx_register_instructions"]; ?>
             
         <form name='register_form' action='<?php echo BASEURL; ?>index.php?page=register' method='post'>    
         <table>
         <tr><td>
-            <?php echo $h->lang["user_signin_register_username"]; ?>&nbsp; </td><td><input type='text' size=30 name='username' value='<?php echo $username_check; ?>' />
+            <?php echo $h->lang["user_signin_register_username"]; ?>&nbsp; </td>
+            <td><input type='text' size=30 name='username' value='<?php echo $h->vars['rpx_profile']['preferredUsername']; ?>' />
             <br /><small><?php echo $h->lang["user_signin_register_username_error_short"]; ?></small>
         </td></tr>
         
         <tr><td>
-            <?php echo $h->lang["user_signin_register_email"]; ?>&nbsp; </td><td><input type='text' size=30 name='email' value='<?php echo $email_check; ?>' />
-        </td></tr>
-        
-        <tr><td>
-            <?php echo $h->lang["user_signin_register_password"]; ?>&nbsp; </td><td><input type='password' size=30 name='password' value='<?php echo $password_check; ?>' />
-            <br /><small><?php echo $h->lang["user_signin_register_password_error_short"]; ?></small>
-        </td></tr>
-        
-        <tr><td>
-            <?php echo $h->lang["user_signin_register_password_verify"]; ?>&nbsp; </td><td><input type='password' size=30 name='password2' value='<?php echo $password2_check; ?>' />
+            <?php echo $h->lang["user_signin_register_email"]; ?>&nbsp; </td>
+            <td><input type='text' size=30 name='email' value='<?php echo $h->vars['rpx_profile']['email']; ?>' />
         </td></tr>
         
         <?php 
             if ($h->vars['useRecaptcha']) { 
                 $user_signin_settings = $h->getSerializedSettings('user_signin');
                 $recaptcha_pubkey = $user_signin_settings['recaptcha_pubkey'];
+                echo "<tr><td colspan=2>" . $h->lang["rpx_register_captcha"] . "</td></tr>";
                 echo "<tr><td colspan=2>" . recaptcha_get_html($recaptcha_pubkey) . "</td></tr>";
             }
         ?>
         
         <input type='hidden' name='users_type' value='register' />
+        <input type='hidden' name='rpx' value='true' />
+        <input type='hidden' name='identifier' value='<?php echo $h->vars['rpx_profile']['identifier']; ?>' />
         <input type='hidden' name='csrf' value='<?php echo $h->csrfToken; ?>' />
         <tr><td>&nbsp;</td><td style='text-align:right;'><input type='submit' class='submit' value='<?php echo $h->lang['user_signin_register_form_submit']; ?>' /></td></tr>            
         </table>
         </form>
     </div>
+    
+    <?php } else { $h->showMessage(); } ?>
